Struct s2n_tls::connection::Connection

pub struct Connection { /* private fields */ }

Implementations

impl Connection

pub fn new(mode: Mode) -> Self

pub fn new_client() -> Self

pub fn new_server() -> Self

pub fn set_blinding(&mut self, blinding: Blinding) -> Result<&mut Self, Error>

can be used to configure s2n to either use built-in blinding (set blinding to Blinding::BuiltIn) or self-service blinding (set blinding to Blinding::SelfService).

pub fn remaining_blinding_delay(&self) -> Result<Duration, Error>

Reports the remaining nanoseconds before the connection may be gracefully shutdown.

This method is expected to succeed, but could fail if the underlying C call encounters errors. Failure indicates that calls to Self::poll_shutdown will also fail and that a graceful two-way shutdown of the connection will not be possible.

pub fn set_client_auth_type(
&mut self,
client_auth_type: ClientAuthType
) -> Result<&mut Self, Error>

Sets whether or not a Client Certificate should be required to complete the TLS Connection.

If this is set to ClientAuthType::Optional the server will request a client certificate but allow the client to not provide one. Rejecting a client certificate when using ClientAuthType::Optional will terminate the handshake.

pub fn set_config(&mut self, config: Config) -> Result<&mut Self, Error>

Associates a configuration object with a connection.

pub fn set_security_policy(
&mut self,
policy: &Policy
) -> Result<&mut Self, Error>

pub fn set_dynamic_record_threshold(
&mut self,
resize_threshold: u32,
timeout_threshold: u16
) -> Result<&mut Self, Error>

provides a smooth transition from s2n_connection_prefer_low_latency to s2n_connection_prefer_throughput.

s2n_send uses small TLS records that fit into a single TCP segment for the resize_threshold bytes (cap to 8M) of data and reset record size back to a single segment after timeout_threshold seconds of inactivity.

pub fn set_application_protocol_preference<P: IntoIterator<Item = I>, I: AsRef<[u8]>>(
&mut self,
protocols: P
) -> Result<&mut Self, Error>

sets the application protocol preferences on an s2n_connection object.

protocols is a list in order of preference, with most preferred protocol first, and of length protocol_count. When acting as a client the protocol list is included in the Client Hello message as the ALPN extension. As a server, the list is used to negotiate a mutual application protocol with the client. After the negotiation for the connection has completed, the agreed upon protocol can be retrieved with s2n_get_application_protocol

pub fn append_application_protocol_preference(
&mut self,
protocol: &[u8]
) -> Result<&mut Self, Error>

pub fn set_receive_callback(
&mut self,
callback: s2n_recv_fn
) -> Result<&mut Self, Error>

may be used to receive data with callbacks defined by the user.

pub unsafe fn set_receive_context(
&mut self,
context: *mut c_void
) -> Result<&mut Self, Error>

Safety

The context pointer must live at least as long as the connection

pub fn set_send_callback(
&mut self,
callback: s2n_send_fn
) -> Result<&mut Self, Error>

may be used to receive data with callbacks defined by the user.

pub unsafe fn set_send_context(
&mut self,
context: *mut c_void
) -> Result<&mut Self, Error>

Safety

The context pointer must live at least as long as the connection

pub fn prefer_low_latency(&mut self) -> Result<&mut Self, Error>

Connections prefering low latency will be encrypted using small record sizes that can be decrypted sooner by the recipient.

pub fn prefer_throughput(&mut self) -> Result<&mut Self, Error>

Connections prefering throughput will use large record sizes that minimize overhead.

pub fn release_buffers(&mut self) -> Result<&mut Self, Error>

wipes and free the in and out buffers associated with a connection.

This function may be called when a connection is in keep-alive or idle state to reduce memory overhead of long lived connections.

pub fn use_corked_io(&mut self) -> Result<&mut Self, Error>

pub fn wipe(&mut self) -> Result<&mut Self, Error>

wipes an existing connection and allows it to be reused.

This method erases all data associated with a connection including pending reads. This function should be called after all I/O is completed and s2n_shutdown has been called. Reusing the same connection handle(s) is more performant than repeatedly calling s2n_connection_new and s2n_connection_free

pub fn poll_negotiate(&mut self) -> Poll<Result<&mut Self, Error>>

Performs the TLS handshake to completion

pub fn poll_send(&mut self, buf: &[u8]) -> Poll<Result<usize, Error>>

Encrypts and sends data on a connection where negotiate has succeeded.

Returns the number of bytes written, and may indicate a partial write.

pub fn poll_recv(&mut self, buf: &mut [u8]) -> Poll<Result<usize, Error>>

Reads and decrypts data from a connection where negotiate has succeeded.

Returns the number of bytes read, and may indicate a partial read. 0 bytes returned indicates EOF due to connection closure.

pub fn poll_flush(&mut self) -> Poll<Result<&mut Self, Error>>

Attempts to flush any data previously buffered by a call to send.

pub fn peek_len(&self) -> usize

Gets the number of bytes that are currently available in the buffer to be read.

pub fn poll_shutdown(&mut self) -> Poll<Result<&mut Self, Error>>

Attempts a graceful shutdown of the TLS connection.

The shutdown is not complete until the necessary shutdown messages have been successfully sent and received. If the peer does not respond correctly, the graceful shutdown may fail.

pub fn alert(&self) -> Option<u8>

Returns the TLS alert code, if any

pub fn set_server_name(&mut self, server_name: &str) -> Result<&mut Self, Error>

Sets the server name value for the connection

pub fn server_name(&self) -> Option<&str>

Get the server name associated with the connection client hello.

pub fn set_waker(&mut self, waker: Option<&Waker>) -> Result<&mut Self, Error>

Sets a Waker on the connection context or clears it if None is passed.

pub fn waker(&self) -> Option<&Waker>

Returns the Waker set on the connection context.

pub fn server_name_extension_used(&mut self)

Mark that the server_name extension was used to configure the connection.

pub fn client_cert_used(&self) -> bool

Check if client auth was used for a connection.

This is only relevant if `ClientAuthType::Optional was used.

pub fn client_cert_chain_bytes(&self) -> Result<Option<&[u8]>, Error>

Retrieves the raw bytes of the client cert chain received from the peer, if present.

pub fn actual_protocol_version(&self) -> Result<Version, Error>

pub fn handshake_type(&self) -> Result<&str, Error>

pub fn cipher_suite(&self) -> Result<&str, Error>

impl Connection

pub fn enable_quic(&mut self) -> Result<&mut Self, Error>

pub fn set_quic_transport_parameters(
&mut self,
buffer: &[u8]
) -> Result<&mut Self, Error>

pub fn quic_transport_parameters(&mut self) -> Result<&[u8], Error>

pub unsafe fn set_secret_callback(
&mut self,
callback: s2n_secret_cb,
context: *mut c_void
) -> Result<&mut Self, Error>

Safety

The context pointer must live at least as long as the connection